arabic_trnsl/arabic_trnsl/requirement.txt
django-rest-frameword
django-cors-headers

pyaudio
sacremoses
sentencepiece
setuptools
SpeechRecognition
torch
torchvision
torchaudio
transformers
vosk
